ci: modify release job to run on non-tag pushes
Change the release job condition to execute when the GitHub ref does not start
with 'refs/tags/' in order to trigger on pushes that are not tags. This
ensures that release-related steps are not run unnecessarily during tag
creation.

Also, add Python version classifiers in pyproject.toml to specify the supported
Python versions for the project.